add field exposing capability
