+++ src/formatters/index.js
@@ -1,5 +1,7 @@
 const graylog = require('./graylog');
+const prettyPrint = require('./prettyPrint');
 
 module.exports = {
   graylog,
+  prettyPrint,
 };
