+++ src/logger.js
@@ -1,34 +1,12 @@
-const _ = require('lodash');
 const domain = require('domain');
-const prettyjson = require('prettyjson');
 const stringify = require('json-stringify-safe');
-
+const { prettyPrint } = require('./formatters');
 const eventFormatter = require('./event-formatter');
 const isEnabled = require('./is-enabled');
 const Serializer = require('./serializer');
 const loggerLevels = require('./levels');
 const logLevelFilter = require('./log-level-filter');
 
-/**
- * Default options for prettyjson
- */
-const PrettyJsonDefaultOptions = Object.freeze({
-  defaultIndentation: 4,
-  inlineArrays: 1,
-});
-
-/**
- * For development mode
- */
-const prettyPrint = (event) => {
-  const header = `[${event.timestamp}]: [${event.message}]`;
-  const eventDetails = _.omit(event, 'timestamp', 'message');
-  const body = prettyjson
-    .render(eventDetails, _.clone(PrettyJsonDefaultOptions))
-    .replace(/\n/g, '\n\t');
-
-  return `\n${header}\n\t${body}\n\n`;
-};
 
 /**
  * default values for Logger instance
@@ -88,6 +66,9 @@ class Logger {
    * You can also exclude specific debuggers by prefixing them with a "-" character
    * `logPatterns: 'api,-api:myAwesomeApi'`
    * @param {String} options.namespace - Logger namespace
+   * @param {Array} [options.fieldsToExpose=undefined] - List of fields to extract from any level
+   * and expose then on first level of log object. Can also be renamed by passing an alias value:
+   * [{fieldName:'level'}, {fieldName:'correlationId', alias:correlation_id}]
    *
    * @param  {...any} extraParameters - DEPRECATED, Prefer usage of options object
    *
@@ -102,12 +83,13 @@ class Logger {
   // eslint-disable-next-line max-lines-per-function
   constructor(options, ...extraParameters) {
     const {
-      context, namespace, logFormat, logPatterns, logLimit, logLevel,
+      context, namespace, logFormat, logPatterns, logLimit, logLevel, fieldsToExpose,
     } = normalizeArguments(options, extraParameters);
 
     Object.assign(this, {
       contextData: { context, name: process.env.APP_NAME },
-      format: process.env.LOGS_PRETTY_PRINT ? prettyPrint : stringify,
+      fieldsToExpose,
+      format: process.env.LOGS_PRETTY_PRINT === '1' ? prettyPrint : stringify,
       log: this.info,
       logFormat,
       logLevel,
@@ -198,7 +180,7 @@ class Logger {
       message, additionalArguments, outputType,
     );
 
-    const formattedEvent = eventFormatter(event, this.logFormat);
+    const formattedEvent = eventFormatter(event, this.fieldsToExpose, this.logFormat);
 
     console.log(`${this.format(formattedEvent)}`); // eslint-disable-line no-console
   }
